Update the controller for events and all that jazz
