+++ nodeapi/api/routes/events.js
@@ -4,127 +4,16 @@ const mongoose = require('mongoose');
 const Event = require('../models/event');
 const Artist = require('../models/artist');
 const checkAuth =  require('../middleware/checkauth');
+const EventController = require('../controllers/events')
 
-router.get('/', (req, res, next) => {
-  Event.find()
-    .select('_id artist_ids eventdatetime venuename venuephone eventlinks')
-    .populate('artist_ids', 'artistname artistpicture')
-    .exec()
-    .then(docs =>{
-      const response = {
-        count: docs.length,
-        Event: docs.map(doc=>{
-          return{
-            _id: doc._id,
-            artist_ids: doc.artist_ids,
-            eventdatetime: doc.eventdatetime,
-            venuename: doc.venuename,
-            venuephone: doc.venuephone,
-            eventlinks: doc.eventlinks,
-            request: {
-              type: 'GET PATCH DELETE',
-              eventurl: 'http://'+req.headers.host+'/events'+doc._id
-            }
-          }
-        })
-      };
-      (docs.length > 0) ? res.status(200).json(response) : res.status(204).json({message: 'No events currently listed'});
-    })
-    .catch( err => {
-      console.log(err);
-      res.status(500).json({error: err});
-    });
-});
+router.get('/', EventController.get_all_events);
 
-router.post('/', checkAuth, (req, res, next) => {
-  let eventdatetime = req.body.eventdatetime;
-  let venuename = req.body.venuename;
-  let venuephone = req.body.venuephone;
-  let eventlinks = req.body.eventlinks;
-  let artist_ids = req.body.artist_ids;
-  Artist.findById(artist_ids)
-    .then(artist => {
-      if (!artist) {
-        return res.status(404).json({
-          message: "Artist not found"
-        });
-      }
-      let event = new Event({
-        _id: new mongoose.Types.ObjectID(),
-        artist_ids: artist_ids,
-        venuename: venuename,
-        venuephone: venuephone,
-        eventlinks: eventlinks,
-        eventdatetime: eventdatetime
-      });
-      return event.save();
-    })
-    .then(result => {
-      res.status(201).json({
-        message: "Event Creation Successful",
-        createdEvent: {
-          _id: result._id,
-          artist_ids: result.artist_ids,
-          venuename: result.venuename,
-          venuephone: result.venuephone,
-          eventlinks: result.eventlinks,
-          eventdatetime: results.eventdatetime
-        }
-      })
-    })
-    .catch(err => {
-      console.log(err);
-      res.status(500).json({
-        error: err
-      });
-    });
-});
+router.post('/', checkAuth, EventController.create_new_event);
 
-router.get('/:eventID', (req, res, next) => {
-  let id = req.params.eventID;
-  Event.findbyId(id)
-    .select('_id artist_ids venuename venuephone eventlinks eventdatetime')
-    .populate('artist_ids', '_id artistname artistpicture')
-    .exec()
-    .then(doc => {
-      (doc) ? res.status(200).json(doc) : res.status(404).json({message: "Event not found", event_id: id});
-    })
-    .catch(err =>{
-      console.log(err);
-      res.status(500).json({error: err});
-    });
-});
+router.get('/:eventID', EventController.get_single_event);
 
-router.patch('/:eventID', checkAuth, (req, res, next) => {
-  let id = req.params.eventID;
-  let updateOps = {};
-  for(let ops of req.body){
-    updateOps[ops.propName] = ops.value;
-  }
-  Event.update({_id: id}, { $set: updateOps })
-    .select('_id artist_ids venuename venuephone eventlinks eventdatetime')
-    .exec()
-    .then(result => {
-      console.log(result);
-      res.status(200).json(result);
-    })
-    .catch(err => {
-      console.log(err);
-      res.status(500).json({error: err});
-    });
-});
+router.patch('/:eventID', checkAuth, EventController.update_event);
 
-router.delete('/:eventID', checkAuth, (req, res, next) => {
-  let id = req.params.eventID;
-  Event.deleteOne({_id: id})
-    .exec()
-    .then(result => {
-      res.status(200).json(result);
-    })
-    .catch(err => {
-      console.log(err);
-      res.status(500).json({error: err});
-    });
-});
+router.delete('/:eventID', checkAuth, EventController.delete_event);
 
 module.exports = router;
